feat: implement checkGameIsOver refactor: moved functions to separate files
