+++ src/components/Game/Game.tsx
@@ -5,11 +5,15 @@ import GameField from '../GameField/GameField'
 import './style.scss'
 import { getRandomFieldValue } from '@helpers/getRandomFieldValue'
 import { getInitialData } from '@helpers/getInitialData'
-import { INITIAL_MOVES, KEYS } from '../../constants'
 import { GameDataType } from '../../types/types.'
 import { getCopyOfArray } from '@helpers/getCopyOfArray'
 import _ from 'lodash'
 import { isIdenticalArrays } from '@helpers/isIdenticalArrays'
+import swipeLeft from '../../utils/swipeLeft'
+import swipeRight from '../../utils/swipeRight'
+import swipeUp from '../../utils/swipeUp'
+import swipeDown from '../../utils/swipeDown'
+import { INITIAL_MOVES, KEYS } from '../../constants'
 
 const Game = (): JSX.Element => {
   const [fieldSize, setFieldSize] = useState<number>(4)
@@ -19,7 +23,9 @@ const Game = (): JSX.Element => {
   const [isPause, setPause] = useState(false)
   const [pauseDelay, setPauseDelay] = useState(0)
   const [startPauseDelay, setStartPauseDelay] = useState<Date | undefined>()
-  const [languageIsEn, setLanguage] = useState(false)
+  const [languageIsEn, setLanguage] = useState(true)
+  const [gameIsOver, setGameOver] = useState(false)
+  const [gameIsStart, setStartGame] = useState(false)
 
   const addRandomValue = (newData: GameDataType) => {
     let isAdded = false
@@ -32,6 +38,11 @@ const Game = (): JSX.Element => {
         newData[getRandomRow][getRandomColumn] = getRandomFieldValue()
         isAdded = true
       }
+
+      if (gameIsStart && !checkGameIsOver(newData)) {
+        console.log('you lose')
+        return
+      }
     }
   }
 
@@ -42,11 +53,11 @@ const Game = (): JSX.Element => {
       addRandomValue(initialData)
     }
 
+    setStartGame(true)
     setGameData(initialData)
   }
 
   const setTimer = () => {
-    console.log(isPause)
     if (!isPause) {
       const diffTime = new Date().getTime() - initTime.getTime() - pauseDelay
       setNowTime(new Date(diffTime))
@@ -54,10 +65,8 @@ const Game = (): JSX.Element => {
   }
 
   useEffect(() => {
+    // check LocalStorage
     initField()
-
-    window.addEventListener('keydown', handleKeyDown)
-    return () => window.removeEventListener('keydown', handleKeyDown)
   }, [])
 
   useEffect(() => {
@@ -65,224 +74,84 @@ const Game = (): JSX.Element => {
     return () => clearInterval(updateTimeInterval)
   }, [isPause, pauseDelay, startPauseDelay])
 
-  const swipeLeft = (copyData: GameDataType) => {
-    const newData = getCopyOfArray(copyData)
+  const handleKeydown = (e: KeyboardEvent) => {
+    let newData = []
 
-    for (let i = 0; i < fieldSize; i += 1) {
-      const row = newData[i]
-      let prevIndex = 0
-      let currentIndex = 1
-
-      while (prevIndex < fieldSize) {
-        if (currentIndex === fieldSize) {
-          currentIndex = prevIndex + 1
-          prevIndex += 1
-          continue
-        }
-        if (row[prevIndex] === 0 && row[currentIndex] === 0) {
-          currentIndex += 1
-        } else if (row[prevIndex] === 0 && row[currentIndex] !== 0) {
-          row[prevIndex] = row[currentIndex]
-          row[currentIndex] = 0
-          currentIndex += 1
-        } else if (row[prevIndex] !== 0 && row[currentIndex] === 0) {
-          currentIndex += 1
-        } else if (row[prevIndex] !== 0 && row[currentIndex] !== 0) {
-          if (row[prevIndex] === row[currentIndex]) {
-            row[prevIndex] = row[prevIndex] + row[currentIndex]
-            row[currentIndex] = 0
-            currentIndex = prevIndex + 1
-            prevIndex += 1
-          } else {
-            prevIndex += 1
-            currentIndex = prevIndex + 1
-          }
-        }
-      }
-    }
-    if (!isIdenticalArrays(copyData, newData)) {
-      addRandomValue(newData)
-    }
-
-    return newData
-  }
-
-  const swipeRight = (copyData: GameDataType) => {
-    const newData = getCopyOfArray(copyData)
-
-    for (let i = fieldSize - 1; i >= 0; i -= 1) {
-      const row = newData[i]
-      let prevIndex = row.length - 1
-      let currentIndex = prevIndex - 1
-
-      while (prevIndex > 0) {
-        if (currentIndex === -1) {
-          currentIndex = prevIndex - 1
-          prevIndex -= 1
-          continue
-        }
-        if (row[prevIndex] === 0 && row[currentIndex] === 0) {
-          currentIndex -= 1
-        } else if (row[prevIndex] === 0 && row[currentIndex] !== 0) {
-          row[prevIndex] = row[currentIndex]
-          row[currentIndex] = 0
-          currentIndex -= 1
-        } else if (row[prevIndex] !== 0 && row[currentIndex] === 0) {
-          currentIndex -= 1
-        } else if (row[prevIndex] !== 0 && row[currentIndex] !== 0) {
-          if (row[prevIndex] === row[currentIndex]) {
-            row[prevIndex] = row[prevIndex] + row[currentIndex]
-            row[currentIndex] = 0
-            currentIndex = prevIndex - 1
-            prevIndex -= 1
-          } else {
-            prevIndex -= 1
-            currentIndex = prevIndex - 1
-          }
-        }
-      }
-    }
-    if (!isIdenticalArrays(copyData, newData)) {
-      addRandomValue(newData)
-    }
-
-    return newData
-  }
-
-  const swipeUp = (copyData: GameDataType) => {
-    const newData = getCopyOfArray(copyData)
-
-    for (let i = 0; i < fieldSize; i += 1) {
-      let prevIndex = 0
-      let currentIndex = 1
-
-      while (prevIndex < fieldSize) {
-        if (currentIndex === fieldSize) {
-          currentIndex = prevIndex + 1
-          prevIndex += 1
-          continue
-        }
-        if (newData[prevIndex][i] === 0 && newData[currentIndex][i] === 0) {
-          currentIndex += 1
-        } else if (newData[prevIndex][i] === 0 && newData[currentIndex][i] !== 0) {
-          newData[prevIndex][i] = newData[currentIndex][i]
-          newData[currentIndex][i] = 0
-          currentIndex += 1
-        } else if (newData[prevIndex][i] !== 0 && newData[currentIndex][i] === 0) {
-          currentIndex += 1
-        } else if (newData[prevIndex][i] !== 0 && newData[currentIndex][i] !== 0) {
-          if (newData[prevIndex][i] === newData[currentIndex][i]) {
-            newData[prevIndex][i] = newData[prevIndex][i] + newData[currentIndex][i]
-            newData[currentIndex][i] = 0
-            currentIndex = prevIndex + 1
-            prevIndex += 1
-          } else {
-            prevIndex += 1
-            currentIndex = prevIndex + 1
-          }
-        }
-      }
-    }
-    if (!isIdenticalArrays(copyData, newData)) {
-      addRandomValue(newData)
-    }
-
-    return newData
-  }
-
-  const swipeDown = (copyData: GameDataType) => {
-    const newData = getCopyOfArray(copyData)
-
-    for (let i = fieldSize - 1; i >= 0; i -= 1) {
-      let prevIndex = newData.length - 1
-      let currentIndex = prevIndex - 1
-
-      while (prevIndex > 0) {
-        if (currentIndex === -1) {
-          currentIndex = prevIndex - 1
-          prevIndex -= 1
-          continue
-        }
-        if (newData[prevIndex][i] === 0 && newData[currentIndex][i] === 0) {
-          currentIndex -= 1
-        } else if (newData[prevIndex][i] === 0 && newData[currentIndex][i] !== 0) {
-          newData[prevIndex][i] = newData[currentIndex][i]
-          newData[currentIndex][i] = 0
-          currentIndex -= 1
-        } else if (newData[prevIndex][i] !== 0 && newData[currentIndex][i] === 0) {
-          currentIndex -= 1
-        } else if (newData[prevIndex][i] !== 0 && newData[currentIndex][i] !== 0) {
-          if (newData[prevIndex][i] === newData[currentIndex][i]) {
-            newData[prevIndex][i] = newData[prevIndex][i] + newData[currentIndex][i]
-            newData[currentIndex][i] = 0
-            currentIndex = prevIndex - 1
-            prevIndex -= 1
-          } else {
-            prevIndex -= 1
-            currentIndex = prevIndex - 1
-          }
-        }
-      }
-    }
-    if (!isIdenticalArrays(copyData, newData)) {
-      addRandomValue(newData)
-    }
-
-    return newData
-  }
-
-  const handleKeyDown = (e: KeyboardEvent) => {
     switch (e.code) {
       case KEYS.LEFT:
       case KEYS.KEY_A:
-        setGameData(prevState => {
-          const copyData = getCopyOfArray(prevState)
-          return swipeLeft(copyData)
-        })
+        newData = swipeLeft(gameData, fieldSize)
+        if (!isIdenticalArrays(gameData, newData)) {
+          addRandomValue(newData)
+        }
+        setGameData(newData)
         break
       case KEYS.RIGHT:
       case KEYS.KEY_D:
-        setGameData(prevState => {
-          const copyData = getCopyOfArray(prevState)
-          return swipeRight(copyData)
-        })
+        newData = swipeRight(gameData, fieldSize)
+        if (!isIdenticalArrays(gameData, newData)) {
+          addRandomValue(newData)
+        }
+        setGameData(newData)
         break
       case KEYS.ARROW_UP:
       case KEYS.KEY_W:
-        setGameData(prevState => {
-          const copyData = getCopyOfArray(prevState)
-          return swipeUp(copyData)
-        })
+        newData = swipeUp(gameData, fieldSize)
+        if (!isIdenticalArrays(gameData, newData)) {
+          addRandomValue(newData)
+        }
+        setGameData(newData)
         break
       case KEYS.ARROW_DOWN:
       case KEYS.KEY_S:
-        setGameData(prevState => {
-          const copyData = getCopyOfArray(prevState)
-          return swipeDown(copyData)
-        })
+        newData = swipeDown(gameData, fieldSize)
+        if (!isIdenticalArrays(gameData, newData)) {
+          addRandomValue(newData)
+        }
+        setGameData(newData)
         break
       default:
         break
     }
   }
 
+  useEffect(() => {
+    window.addEventListener('keydown', handleKeydown)
+    return () => window.removeEventListener('keydown', handleKeydown)
+  })
+
   const onSetPause = () => {
     if (!isPause) {
       setStartPauseDelay(new Date())
-      setPauseDelay(0)
       setPause(!isPause)
     } else {
       const pause = new Date().getTime() - startPauseDelay!.getTime()
-      setPauseDelay(pause)
+      setPauseDelay(pauseDelay + pause)
       setPause(!isPause)
     }
+  }
+
+  const checkGameIsOver = (newData: GameDataType): boolean => {
+    if (newData.flat().every(cell => cell !== 0)) {
+      for (let i = 0; i < fieldSize; i += 1) {
+        for (let j = 0; j < fieldSize - 1; j += 1) {
+          if (newData[i][j] === newData[i][j + 1]) {
+            return true
+          } else if (newData[j][i] === newData[j + 1][i]) {
+            return true
+          }
+        }
+      }
+      return false
+    }
+    return true
   }
 
   return (
     <div className="game-wrapper">
       <GameMenu
         setGameData={setGameData}
-        addRandomValue={addRandomValue}
+        // addRandomValue={addRandomValue}
         fieldSize={fieldSize}
         onSetPause={onSetPause}
         languageIsEn={languageIsEn}
