+++ src/components/Game/style.scss
@@ -1,6 +1,7 @@
 .game-wrapper {
   display: flex;
   flex-direction: column;
+  max-width: 500px;
 
-  padding: 20px;
+  margin: 20px;
 }
